+++ src/migrations/Install.php
@@ -14,7 +14,7 @@ use putyourlightson\blitz\records\ElementCacheRecord;
 use putyourlightson\blitz\records\ElementExpiryDateRecord;
 use putyourlightson\blitz\records\ElementQueryCacheRecord;
 use putyourlightson\blitz\records\ElementQueryRecord;
-use putyourlightson\blitz\records\CacheFlagRecord;
+use putyourlightson\blitz\records\CacheTagRecord;
 
 class Install extends Migration
 {
@@ -47,7 +47,7 @@ class Install extends Migration
         $this->dropTableIfExists(ElementQueryRecord::tableName());
         $this->dropTableIfExists(ElementCacheRecord::tableName());
         $this->dropTableIfExists(ElementExpiryDateRecord::tableName());
-        $this->dropTableIfExists(CacheFlagRecord::tableName());
+        $this->dropTableIfExists(CacheTagRecord::tableName());
         $this->dropTableIfExists(CacheRecord::tableName());
 
         return true;
@@ -102,10 +102,10 @@ class Install extends Migration
             ]);
         }
 
-        if (!$this->db->tableExists(CacheFlagRecord::tableName())) {
-            $this->createTable(CacheFlagRecord::tableName(), [
+        if (!$this->db->tableExists(CacheTagRecord::tableName())) {
+            $this->createTable(CacheTagRecord::tableName(), [
                 'cacheId' => $this->integer()->notNull(),
-                'flag' => $this->string()->notNull(),
+                'tag' => $this->string()->notNull(),
             ]);
         }
 
@@ -126,7 +126,7 @@ class Install extends Migration
         $this->createIndex(null, ElementQueryCacheRecord::tableName(), ['cacheId', 'queryId'], true);
         $this->createIndex(null, ElementQueryRecord::tableName(), 'index', true);
         $this->createIndex(null, ElementQueryRecord::tableName(), 'type', false);
-        $this->createIndex(null, CacheFlagRecord::tableName(), 'flag', false);
+        $this->createIndex(null, CacheTagRecord::tableName(), 'tag', false);
     }
 
     /**
@@ -142,6 +142,6 @@ class Install extends Migration
         $this->addForeignKey(null, ElementExpiryDateRecord::tableName(), 'elementId', Element::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, ElementQueryCacheRecord::tableName(), 'cacheId', CacheRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         $this->addForeignKey(null, ElementQueryCacheRecord::tableName(), 'queryId', ElementQueryRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
-        $this->addForeignKey(null, CacheFlagRecord::tableName(), 'cacheId', CacheRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
+        $this->addForeignKey(null, CacheTagRecord::tableName(), 'cacheId', CacheRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
     }
 }