+++ src/migrations/m190227_120000_add_cachetags_table.php
@@ -5,9 +5,9 @@ namespace putyourlightson\blitz\migrations;
 use Craft;
 use craft\db\Migration;
 use putyourlightson\blitz\records\CacheRecord;
-use putyourlightson\blitz\records\CacheFlagRecord;
+use putyourlightson\blitz\records\CacheTagRecord;
 
-class m190227_120000_add_cacheflags_table extends Migration
+class m190227_120000_add_cachetags_table extends Migration
 {
     // Public Methods
     // =========================================================================
@@ -17,15 +17,15 @@ class m190227_120000_add_cacheflags_table extends Migration
      */
     public function safeUp()
     {
-        $table = CacheFlagRecord::tableName();
+        $table = CacheTagRecord::tableName();
 
         if (!$this->db->tableExists($table)) {
             $this->createTable($table, [
                 'cacheId' => $this->integer()->notNull(),
-                'flag' => $this->string()->notNull(),
+                'tag' => $this->string()->notNull(),
             ]);
 
-            $this->createIndex(null, $table, 'flag', false);
+            $this->createIndex(null, $table, 'tag', false);
 
             $this->addForeignKey(null, $table, 'cacheId', CacheRecord::tableName(), 'id', 'CASCADE', 'CASCADE');
         }